# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la gestión de una peluquería",
        "title": "Peluquería API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/servicio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve una lista de todos los servicios registrados en el sistema.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener todos los servicios",
                "responses": {
                    "200": {
                        "description": "Servicios obtenidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.GetServiceDto"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite crear un nuevo servicio en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear servicio",
                "parameters": [
                    {
                        "description": "Datos del servicio",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ServiceDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio creado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/servicio/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve los datos de un servicio específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener servicio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio obtenido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.GetServiceDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite actualizar los datos de un servicio existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar servicio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del servicio",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ServiceDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio actualizado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID o datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite eliminar un servicio específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar servicio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Detalle del error",
                    "type": "string",
                    "example": "Descripción del error"
                },
                "status": {
                    "description": "Siempre \"error\"",
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "dtos.GetServiceDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Corte de pelo clasico"
                },
                "estimated_time_minutes": {
                    "type": "integer",
                    "example": 90
                },
                "name": {
                    "type": "string",
                    "example": "Corte de pelo"
                },
                "price": {
                    "type": "number",
                    "example": 10000
                }
            }
        },
...
        "dtos.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Datos adicionales"
                },
                "message": {
                    "description": "Mensaje descriptivo",
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "status": {
                    "description": "Ejemplo de estado para éxito o error",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "dtos.ServiceDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Corte de pelo clasico"
                },
                "estimated_time_hours": {
                    "type": "integer",
                    "example": 1
                },
                "estimated_time_minutes": {
                    "type": "integer",
                    "example": 30
                },
                "name": {
                    "type": "string",
                    "example": "Corte de pelo"
                },
                "price": {
                    "type": "number",
                    "example": 10000
                }
            }
        },
...
    }
}